Refactor awaitPersistenceInit
- Move it to a new module: persistence-testkit.
- Deprecate old locations.
