+++ persistence-cassandra/javadsl/src/main/scala/com/lightbend/lagom/javadsl/persistence/cassandra/testkit/TestUtil.scala
@@ -7,6 +7,7 @@ import com.lightbend.lagom.internal.javadsl.persistence.testkit.CassandraTestCon
 import com.lightbend.lagom.javadsl.persistence.testkit.AbstractTestUtil
 import com.typesafe.config.Config
 
+@deprecated("Internal object, not intended for direct use.", "1.5.0")
 object TestUtil extends AbstractTestUtil {
 
   def persistenceConfig(testName: String, cassandraPort: Int): Config = CassandraTestConfig.persistenceConfig(testName, cassandraPort)
