+++ persistence-cassandra/javadsl/src/test/scala/com/lightbend/lagom/javadsl/persistence/cassandra/CassandraPersistenceSpec.scala
@@ -8,8 +8,9 @@ import java.io.File
 import akka.actor.ActorSystem
 import akka.cluster.Cluster
 import akka.persistence.cassandra.testkit.CassandraLauncher
-import com.lightbend.lagom.persistence.{ ActorSystemSpec, PersistenceSpec }
+import com.lightbend.lagom.internal.persistence.testkit.AwaitPersistenceInit.awaitPersistenceInit
 import com.lightbend.lagom.javadsl.persistence.cassandra.testkit.TestUtil
+import com.lightbend.lagom.persistence.{ ActorSystemSpec, PersistenceSpec }
 import com.typesafe.config.{ Config, ConfigFactory }
 
 class CassandraPersistenceSpec(system: ActorSystem) extends ActorSystemSpec(system) {
@@ -29,7 +30,7 @@ class CassandraPersistenceSpec(system: ActorSystem) extends ActorSystemSpec(syst
 
     val cassandraDirectory = new File("target/" + system.name)
     CassandraLauncher.start(cassandraDirectory, "lagom-test-embedded-cassandra.yaml", clean = true, port = 0)
-    TestUtil.awaitPersistenceInit(system)
+    awaitPersistenceInit(system)
 
     // Join ourselves - needed because the Cassandra offset store uses cluster startup task
     val cluster = Cluster(system)
