+++ persistence-cassandra/scaladsl/src/multi-jvm/scala/com/lightbend/lagom/scaladsl/persistence/cassandra/CassandraClusteredPersistentEntitySpec.scala
@@ -8,18 +8,19 @@ import java.io.File
 import akka.actor.ActorSystem
 import akka.persistence.cassandra.testkit.CassandraLauncher
 import akka.stream.{ ActorMaterializer, Materializer }
+import com.lightbend.lagom.internal.persistence.testkit.AwaitPersistenceInit.awaitPersistenceInit
+import com.lightbend.lagom.scaladsl.api.ServiceLocator
+import com.lightbend.lagom.scaladsl.api.ServiceLocator.NoServiceLocator
 import com.lightbend.lagom.scaladsl.persistence.ReadSideProcessor
 import com.lightbend.lagom.scaladsl.persistence.TestEntity.Evt
 import com.lightbend.lagom.scaladsl.persistence.cassandra.testkit.TestUtil
 import com.lightbend.lagom.scaladsl.persistence.multinode.{ AbstractClusteredPersistentEntityConfig, AbstractClusteredPersistentEntitySpec }
+import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
 import com.typesafe.config.Config
-import play.api.{ Configuration, Environment, Mode }
 import play.api.inject.DefaultApplicationLifecycle
+import play.api.{ Configuration, Environment, Mode }
 
 import scala.concurrent.{ ExecutionContext, Future }
-import com.lightbend.lagom.scaladsl.api.ServiceLocator
-import com.lightbend.lagom.scaladsl.api.ServiceLocator.NoServiceLocator
-import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
 
 object CassandraClusteredPersistentEntityConfig extends AbstractClusteredPersistentEntityConfig {
   override def additionalCommonConfig(databasePort: Int): Config =
@@ -38,7 +39,7 @@ class CassandraClusteredPersistentEntitySpec extends AbstractClusteredPersistent
     runOn(node1) {
       val cassandraDirectory = new File("target/" + system.name)
       CassandraLauncher.start(cassandraDirectory, "lagom-test-embedded-cassandra.yaml", clean = true, port = databasePort)
-      TestUtil.awaitPersistenceInit(system)
+      awaitPersistenceInit(system)
     }
     enterBarrier("cassandra-started")
