+++ persistence-cassandra/scaladsl/src/test/scala/com/lightbend/lagom/scaladsl/persistence/cassandra/CassandraPersistenceSpec.scala
@@ -5,10 +5,11 @@ package com.lightbend.lagom.scaladsl.persistence.cassandra
 
 import java.io.File
 
-import akka.actor.{ ActorSystem, BootstrapSetup }
 import akka.actor.setup.ActorSystemSetup
+import akka.actor.{ ActorSystem, BootstrapSetup }
 import akka.cluster.Cluster
 import akka.persistence.cassandra.testkit.CassandraLauncher
+import com.lightbend.lagom.internal.persistence.testkit.AwaitPersistenceInit.awaitPersistenceInit
 import com.lightbend.lagom.persistence.{ ActorSystemSpec, PersistenceSpec }
 import com.lightbend.lagom.scaladsl.persistence.cassandra.testkit.TestUtil
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializerRegistry
@@ -34,7 +35,7 @@ class CassandraPersistenceSpec private (system: ActorSystem) extends ActorSystem
 
     val cassandraDirectory = new File("target/" + system.name)
     CassandraLauncher.start(cassandraDirectory, "lagom-test-embedded-cassandra.yaml", clean = true, port = 0)
-    TestUtil.awaitPersistenceInit(system)
+    awaitPersistenceInit(system)
 
     // Join ourselves - needed because the Cassandra offset store uses cluster startup task
     val cluster = Cluster(system)
