+++ persistence-jdbc/scaladsl/src/main/scala/com/lightbend/lagom/scaladsl/persistence/jdbc/testkit/TestUtil.scala
@@ -5,4 +5,5 @@ package com.lightbend.lagom.scaladsl.persistence.jdbc.testkit
 
 import com.lightbend.lagom.scaladsl.persistence.testkit.AbstractTestUtil
 
+@deprecated("Internal object, not intended for direct use.", "1.5.0")
 object TestUtil extends AbstractTestUtil
