+++ persistence-jdbc/scaladsl/src/test/scala/com/lightbend/lagom/scaladsl/persistence/slick/SlickPersistenceSpec.scala
@@ -8,6 +8,7 @@ import akka.actor.{ ActorSystem, BootstrapSetup }
 import akka.cluster.Cluster
 import com.lightbend.lagom.internal.persistence.ReadSideConfig
 import com.lightbend.lagom.internal.persistence.jdbc.{ SlickDbTestProvider, SlickOffsetStore, SlickProvider }
+import com.lightbend.lagom.internal.persistence.testkit.AwaitPersistenceInit.awaitPersistenceInit
 import com.lightbend.lagom.internal.scaladsl.persistence.jdbc.OffsetTableConfiguration
 import com.lightbend.lagom.internal.scaladsl.persistence.slick.SlickReadSideImpl
 import com.lightbend.lagom.persistence.{ ActorSystemSpec, PersistenceSpec }
@@ -17,8 +18,8 @@ import com.typesafe.config.{ Config, ConfigFactory }
 import play.api.inject.{ ApplicationLifecycle, DefaultApplicationLifecycle }
 import play.api.{ Configuration, Environment }
 
+import scala.concurrent.Await
 import scala.concurrent.duration._
-import scala.concurrent.{ Await, ExecutionContext }
 
 abstract class SlickPersistenceSpec private (_system: ActorSystem) extends ActorSystemSpec(_system) {
 
@@ -64,7 +65,7 @@ abstract class SlickPersistenceSpec private (_system: ActorSystem) extends Actor
     // Trigger tables to be created
     Await.ready(slick.ensureTablesCreated(), 20.seconds)
 
-    TestUtil.awaitPersistenceInit(system)
+    awaitPersistenceInit(system)
   }
 
   override def afterAll(): Unit = {
